Replaced references to TabLayout with TabView
